requirements.txt
fastapi
uvicorn
langchain
openai
duckduckgo-search
langchain-community
langchain-openai
python-dotenv